+++ docsig/_main.py
@@ -25,6 +25,7 @@ def main() -> int:
         check_dunders=parser.args.check_dunders,
         check_overridden=parser.args.check_overridden,
         check_protected=parser.args.check_protected,
+        summary=parser.args.summary,
         targets=parser.args.target,
         disable=parser.args.disable
     )